% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_b.R
\name{aov_b}
\alias{aov_b}
\title{Analysis of Variance using Bayesian methods}
\usage{
aov_b(
  formula,
  data,
  heteroscedastic = TRUE,
  prior_mean_mu,
  prior_mean_nu = 0.001,
  prior_var_shape = 0.001,
  prior_var_rate = 0.001,
  CI_level = 0.95,
  ROPE = 0.1,
  contrasts,
  improper = FALSE,
  seed = 1,
  mc_error = 0.005
)
}
\arguments{
\item{formula}{A formula specifying the model.}

\item{data}{A data frame in which the variables specified in the formula
will be found. If missing, the variables are searched for in the standard way.}

\item{prior_mean_mu}{numeric. Hyperparameter for the a priori mean of the
group means.}

\item{prior_mean_nu}{numeric. Hyperparameter which scales the precision of
the group means.}

\item{prior_var_shape}{numeric. Twice the shape parameter for the inverse gamma prior on
the residual variance(s).  I.e., \eqn{\sigma^2\sim IG}(prior_var_shape/2,prior_var_rate/2).}

\item{prior_var_rate}{numeric. Twice the rate parameter for the inverse gamma prior on
the residual variance(s).  I.e., \eqn{\sigma^2\sim IG}(prior_var_shape/2,prior_var_rate/2).}

\item{CI_level}{numeric. Credible interval level.}

\item{ROPE}{numeric.  Used to compute posterior probability that Cohen's D +/- ROPE}

\item{contrasts}{numeric/matrix. Either vector of length equal to the number of
levels in the grouping variable, or else a matrix where each row is a separate
contrast.}

\item{improper}{logical.  Should we use an improper prior that is proportional
to the inverse of the variance?}

\item{seed}{integer.  Always set your seed!!!}

\item{mc_error}{The number of posterior draws will ensure that with 99\%
probability the bounds of the credible intervals will be within \eqn{\pm}
\code{mc_error}\eqn{\times 4s_y}, that is, within 100\code{mc_error}\% of the
trimmed range of y.}

\item{heterscedastic}{logical.  Set to FALSE to assume all groups have
equal variance.}
}
\value{
Object of class "aov_b" with the following elements:
\itemize{
\item summary - tibble giving the summary of the model parameters
\item pairwise summary - tibble giving the summary comparing all
factor level means
\item contrasts (if provided) - list with named elements L (the contrasts provided
by the user) and summary.
\item posterior_draws - mcmc object (see coda package) giving the posterior draws
\item formula, data - input by user
\item posterior_parameters -
\itemize{
\item mu_g - the post. means of the group means
\item nu_g - the post. scalars of the precision
\item a_g - (twice) the post. shape of the inv. gamma for the group variances
\item b_g - (twice) the post. rate of the inv. gamma for the group variances.
}
\item hyperparameters -
\itemize{
\item mu - the prior mean of the group means
\item nu - the prior scalar of the precision
\item a - (twice) the prior shape of the inv. gamma for the group variances
\item b - (twice) the prior rate of the inv. gamma for the group variances.
}
}
}
\description{
Analysis of Variance using Bayesian methods
}
\details{
\strong{MODEL:}
The likelihood model is given by
\deqn{
 y_{gi} \overset{iid}{\sim} N(\mu_g,\sigma^2_g),
}
(although if \code{heterscedastic} is set to FALSE, \eqn{\sigma^2_g=\sigma^2_h}
\eqn{\forall g,h}).

The prior is given by
\deqn{
 \mu_g|\sigma^2_g \overset{iid}{\sim} N\left(\mu,\frac{\sigma^2_g}{\nu}\right), \\
 \sigma^2_g \overset{iid}{\sim} \Gamma^{-1}(a/2,b/2),
}
where \eqn{mu} is set by \code{prior_mean_mu}, \eqn{nu} is set by
\code{prior_mean_nu}, \eqn{a} is set by \code{prior_var_shape}, and
\eqn{b} is set by \code{prior_var_rate}.

The posterior is
\deqn{
 \mu_g|y,\sigma^2_g \overset{iid}{\sim} N\left(\hat\mu_g,\frac{\sigma^2_g}{\nu_g}\right), \\
 \sigma^2_g|y \overset{iid}{\sim} \Gamma^{-1}(a_g/2,b_g/2),
}
where \eqn{\hat\mu_g}, \eqn{\nu_g}, \eqn{a_g}, and \eqn{b_g} are all returned
by \code{aov_b} in the named element \code{posterior_parameters}.

\strong{ROPE:}

If missing, the ROPE bounds will be given under the principle of "half of a
small effect size."  Using Cohen's D of 0.2 as a small effect size, the ROPE
is defined in terms of \eqn{-0.1 <} Cohen's D \eqn{ < 0.1}.
}
\references{
Charles R. Doss, James M. Flegal, Galin L. Jones, Ronald C. Neath "Markov chain Monte Carlo estimation of quantiles," Electronic Journal of Statistics, Electron. J. Statist. 8(2), 2448-2478, (2014)
}
