% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDratio.R
\name{BF}
\alias{BF}
\alias{SDratio}
\alias{SDratio.glm_b}
\title{Bayes factors for lm_b and glm_b}
\usage{
SDratio(object, ...)

\method{SDratio}{glm_b}(object, by = "coefficient", ...)
}
\arguments{
\item{object}{lm_b or glm_b object}

\item{by}{character.  Either "coefficient" or "variable".  If the former,
Bayes factors will be computed for each regression coefficient separately.
If the latter, Bayes factors will be computed for each covariate separately.}
}
\description{
Bayes factors for Bayesian regression objects using the Savage-Dickey ratio
}
\details{
Bayes factors are given in terms of favoring the two-tailed alternative hypothesis
vs. the null hypothesis that the regression coefficient equals zero.
Currently implemented for \code{lm_b} or \code{glm_b} objects.  Note
that for \code{glm_b} objects, if importance sampling was used,
the model will be refit using fixed form variational Bayes to get
the multivariate posterior density.

Interpretation is taken from Kass and Raftery.
}
\references{
James M. Dickey. "The Weighted Likelihood Ratio, Linear Hypotheses on Normal Location Parameters." Ann. Math. Statist. 42 (1) 204 - 223, February, 1971. https://doi.org/10.1214/aoms/1177693507

Kass, R. E., & Raftery, A. E. (1995). Bayes Factors. Journal of the American Statistical Association, 90(430), 773–795.
}
