% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma_inference.R
\name{bma_inference}
\alias{bma_inference}
\title{Bayesian model averaging}
\usage{
bma_inference(
  formula,
  data,
  zellner_g = nrow(data),
  CI_level = 0.95,
  mc_draws = 10000,
  seed = 1,
  ...
)
}
\arguments{
\item{formula}{A formula specifying the model.}

\item{data}{Data used in linear regression model}

\item{zellner_g}{numeric.  Positive number giving the value of "g" in Zellner's
g prior.}

\item{CI_level}{Level for credible interval}

\item{mc_draws}{Integer. Number of draws in Monte Carlo integration. Passed into \code{\link[BMS]{bms}}}

\item{seed}{Integer. Always set your seed!!!}

\item{...}{Other arguments for \code{\link[BMS]{bms}}.}
}
\value{
A list with the following elements:
\itemize{
\item summary Tibble with point and interval estimates
\item lm_b_fits A list of lm_b fits using zellner's g prior for
all the top models from \code{\link[BMS]{bms}}
\item hyperparameters A named list with the user-specified zellner's g value.
\item posterior_draws matrix of posterior draws of the regression parameters,
marginalizing out the model
}
}
\description{
Estimates and CIs from BMA
}
\details{
\code{bma_inference} leverages the \code{bms} function from its
eponymous R package, and then uses \code{lm_b} to obtain inference
on the regression coefficients for Bayesian model averaging.
}
