% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateANOVATABLE}
\alias{calculateANOVATABLE}
\title{Generate Anova Table}
\usage{
calculateANOVATABLE(df, B1 = NULL, SXX = NULL, SXY = NULL, SYY = NULL)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'x' and 'y' columns}

\item{B1}{(optional) value of B1 if known. Will be calculated from the data if unknown.}

\item{SXX}{(optional) value of Sxx if known. Will be calculated from the data if unknown.}

\item{SXY}{(optional) value of Sxy if known. Will be calculated from the data if unknown.}

\item{SYY}{(optional) value of Syy if known. Will be calculated from the data if unknown.}
}
\value{
the_analysis_of_variance_table_that_summarize_
the_results_of_an_ANOVA_test_displays_the_sources_of_variations_&\emph{D.O.F}&_
sum_of_squares_&\emph{mean_squares}&_F-statistics
}
\description{
Generate Anova Table
}
