% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateB0}
\alias{calculateB0}
\title{Calculate B0}
\usage{
calculateB0(df, B1 = NULL, sxx = NULL, sxy = NULL)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'x' and 'y' columns}

\item{B1}{(optional) value of B1 if known.Will be calculated from the data if unknown.}

\item{sxx, sxy}{(optional) values of Sxx and Sxy if known. Will be calculated from the data if unknown.
sxx and sxy values are not needed if B1 is known.}
}
\value{
the intercept of the regression line
}
\description{
Calculate B0
}
