% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateB1}
\alias{calculateB1}
\title{Calculate B1}
\usage{
calculateB1(df, sxx = NULL, sxy = NULL)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'x' and 'y' columns}

\item{sxx, sxy}{(optional) values of Sxx and Sxy if known. Will be calculated from the data if unknown.}
}
\value{
the slope of the regression line
}
\description{
Calculate B1
}
