% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateCI_B1}
\alias{calculateCI_B1}
\title{calculateCI_B1}
\usage{
calculateCI_B1(
  data,
  known = FALSE,
  sigma,
  alpha,
  B1 = NULL,
  MSE = NULL,
  SXX = NULL
)
}
\arguments{
\item{data}{a dataset of x and y}

\item{known}{a boolean parameter that takes (T/F) as value}

\item{sigma}{the variance of the dataset and it has  default value (NULL)}

\item{alpha}{the value of significance level}

\item{B1}{(optional) value of B1 if known. Will be calculated from the data if unknown.}

\item{MSE}{(optional) value of MSE if known. Will be calculated from the data if unknown.}

\item{SXX}{(optional) value of Sxx if known. Will be calculated from the data if unknown.}
}
\value{
the confidence interval of B1
}
\description{
This function calculate the confidence interval of the estimator B1
in a linear regression model
}
