% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateCI_mean_response}
\alias{calculateCI_mean_response}
\title{Calculating CI for mean response}
\usage{
calculateCI_mean_response(
  df,
  a,
  xo,
  known = FALSE,
  sigma2 = NULL,
  Bo = NULL,
  B1 = NULL,
  MSE = NULL,
  Sxx = NULL
)
}
\arguments{
\item{df}{a dataset of x and y}

\item{a}{significance level (alpha)}

\item{xo}{value of xo to calculate the mean response at}

\item{known}{boolean to indicate if sigma squared is known or not}

\item{sigma2}{value of sigma squared, if known}

\item{Bo}{(optional) value of Bo if known. Will be calculated from the data if unknown.}

\item{B1}{(optional) value of B1 if known. Will be calculated from the data if unknown.}

\item{MSE}{(optional) value of MSE if known. Will be calculated from the data if unknown.}

\item{Sxx}{(optional) value of Sxx if known. Will be calculated from the data if unknown.}
}
\value{
confidence interval of mean response
}
\description{
Calculating CI for mean response
}
