% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateCI_SigmaSquare}
\alias{calculateCI_SigmaSquare}
\title{caculateCI_for sigma square}
\usage{
calculateCI_SigmaSquare(df, alpha, MSE = NULL)
}
\arguments{
\item{df}{a dataset of x and y}

\item{alpha}{the value of significance level}

\item{MSE}{(optional) value of MSE if known. Will be calculated from the data if unknown.}
}
\value{
the confidence interval of sigma squared
}
\description{
This function calculate the confidence interval of the estimator B1
in a linear regression model
}
