% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateCORRELATIONCOEFF}
\alias{calculateCORRELATIONCOEFF}
\title{Calculate r}
\usage{
calculateCORRELATIONCOEFF(df, B1 = NULL, RSQUARED = NULL)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'x' and 'y' columns}

\item{B1}{(optional) value of B1 if known. Will be calculated from the data if unknown.}

\item{RSQUARED}{(optional) value of RSQUARED if known. Will be calculated from the data if unknown.}
}
\value{
the_correlation_coefficient_that_determines_
the_strength_of_the_relationship_between_2_variables
}
\description{
Calculate r
}
