% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateFVALUE}
\alias{calculateFVALUE}
\title{Calculate Anova F-Value}
\usage{
calculateFVALUE(df, MSR = NULL, MSE = NULL)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'x' and 'y' columns}

\item{MSR}{(optional) value of MSR if known. Will be calculated from the data if unknown.}

\item{MSE}{(optional) value of MSE if known. Will be calculated from the data if unknown.}
}
\value{
the division between MSR & MSE
}
\description{
Calculate Anova F-Value
}
