% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateMSE}
\alias{calculateMSE}
\title{Calculate MSE}
\usage{
calculateMSE(df, SSE = NULL)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'x' and 'y' columns}

\item{SSE}{(optional) values of SSE if known. Will be calculated from the data if unknown.}
}
\value{
estimation for sigma square
}
\description{
Calculate MSE
}
