% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateMSR}
\alias{calculateMSR}
\title{Calculate MSR}
\usage{
calculateMSR(df, SSR = NULL)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'x' and 'y' columns}

\item{SSR(optional)}{values of SSR if known. Will be calculated from the data if unknown.}
}
\value{
the mean square regression
}
\description{
Calculate MSR
}
