% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateRSQUARED}
\alias{calculateRSQUARED}
\title{Calculate R Squared}
\usage{
calculateRSQUARED(
  df,
  SST = NULL,
  SSR = NULL,
  B1 = NULL,
  SXX = NULL,
  SXY = NULL
)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'x' and 'y' columns}

\item{SST, SSR}{(optional) values of SST and SSR if known. Will be calculated from the data if unknown.

The following values are not needed if SSR is known.}

\item{B1}{(optional) value of B1 if known. Will be calculated from the data if unknown.}

\item{SXX, SXY}{(optional) values of Sxx and Sxy if known. Will be calculated from the data if unknown.}
}
\value{
the coefficient of determination that is always +ve in range \link{0,1}
}
\description{
Calculate R Squared
}
