% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateSSE}
\alias{calculateSSE}
\title{Calculate SSE}
\usage{
calculateSSE(df, SST = NULL, SSR = NULL)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'x' and 'y' columns}

\item{SST, SSR}{(optional) values of SST and SSR if known. Will be calculated from the data if unknown.}
}
\value{
the_squared_difference_between_the_actual_values_and_
the_predicted_values
}
\description{
Calculate SSE
}
