% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateSSR}
\alias{calculateSSR}
\title{Calculate SRR}
\usage{
calculateSSR(df, B1 = NULL, SXX = NULL, SXY = NULL)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'x' and 'y' columns}

\item{B1}{(optional) value of B1 if known. Will be calculated from the data if unknown.}

\item{SXX, SXY}{(optional) values of Sxx and Sxy if known. Will be calculated from the data if unknown.
sxy value is not needed if B1 is known.}
}
\value{
the percentage of data that is fitted by the regression model
the_squared_difference_between_the_predicted_values_and_the_mean_of_
the_actual_values
}
\description{
Calculate SRR
}
