% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateSST}
\alias{calculateSST}
\title{Calculate SST}
\usage{
calculateSST(df)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'x' and 'y' columns}
}
\value{
the total variability of the dependent variable around its mean
}
\description{
Calculate SST
}
