% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateSXX}
\alias{calculateSXX}
\title{Calculate Sxx}
\usage{
calculateSXX(df)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'x' column}
}
\value{
sum_of_squared_deviations_of_the_independent_variable_from_its_mean
}
\description{
Calculate Sxx
}
