% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateSXY}
\alias{calculateSXY}
\title{Calculate Sxy}
\usage{
calculateSXY(df)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'x' and 'y' columns}
}
\value{
the_sum_of_the_product_of_the_differences_between_
x_values_and_its_mean_and_the_differences_between_
y_values_and_its_mean
}
\description{
Calculate Sxy
}
