% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateSYY}
\alias{calculateSYY}
\title{Calculate Syy}
\usage{
calculateSYY(df)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'y' column}
}
\value{
sum_of_squared_deviations_of_the_dependent_variable_from_its_mean
}
\description{
Calculate Syy
}
