% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{calculateyo}
\alias{calculateyo}
\title{Calculating y_predicted}
\usage{
calculateyo(df, xo, Bo = NULL, B1 = NULL)
}
\arguments{
\item{df}{a dataset of x and y}

\item{Bo}{(optional) value of Bo if known. Will be calculated from the data if unknown.}

\item{B1}{(optional) value of B1 if known. Will be calculated from the data if unknown.}

\item{x0}{value of x0 to calculate the mean response at}
}
\value{
y0
}
\description{
Calculating y_predicted
}
