% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_control_b.R
\name{case_control_b}
\alias{case_control_b}
\title{Case-Control Analysis}
\usage{
case_control_b(
  cases,
  controls,
  x,
  large_sample_approx,
  prior_mean = 0,
  prior_sd = log(10)/1.96,
  plot = TRUE,
  CI_level = 0.95,
  seed = 1,
  mc_relative_error = 0.01
)
}
\arguments{
\item{cases}{vector of length 2, giving the numbers at risk and not at risk,
respectively, for cases}

\item{controls}{vector of length 2, giving the numbers at risk and not at risk,
respectively, for controls}

\item{x}{2x2 contingency table.  The rows should depict the at risk status
(first row is at risk, second row is not at risk), and the columns should
depict the case control status (first column is case, second column is control).}

\item{large_sample_approx}{If all cell counts of \code{x} are not too low
(\eqn{\geq 5}) then use the approximation that the empirical log odds are
normally distributed.  (See details for more.)  If missing, this will be
set to \code{TRUE} iff all cell counts are greater than or equal to 5.}

\item{prior_mean}{numeric.  The prior mean on the log odds ratio}

\item{prior_sd}{numeric.  The prior sd on the log odds ratio. See details
for default values.}

\item{plot}{logical.  Should a plot be shown?}

\item{CI_level}{The posterior probability to be contained in the
credible interval.}

\item{seed}{integer.  Always set your seed!!! (ignored if \code{large_sample_approx = TRUE}.)}

\item{mc_relative_error}{The relative monte carlo error of the quantiles of the CIs.
(ignored if \code{large_sample_approx = TRUE}.)}

\item{ROPE}{ROPE for odds ratio. Provide either a single value or a vector
of length two.  If the former, the ROPE will be taken as (1/ROPE,ROPE).
If the latter, these will be the bounds of the ROPE.}
}
\description{
Bayesian analysis of a case-control study (without covariates).
}
