% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_test_b.R
\name{chisq_test_b}
\alias{chisq_test_b}
\alias{independence_b}
\title{Test of independence for 2-way contingency tables}
\usage{
independence_b(
  x,
  sampling_design = "multinomial",
  ROPE,
  prior = "jeffreys",
  prior_shapes,
  CI_level = 0.95,
  seed = 1,
  mc_error = 0.001
)
}
\arguments{
\item{x}{Either a table or a matrix of counts}

\item{sampling_design}{Either "multinomial", "fixed rows", or "fixed columns"}

\item{prior}{Either "jeffreys" (Dirichlet(1/2)) or "uniform" (Dirichlet(1)).
This is ignored if prior_shapes is provided.}

\item{prior_shapes}{Either a single positive scalar, in which case a
symmetric Dirichlet is used, or else a matrix matching the dimensions
of x or a vector of length \code{prod(dim(x))}.}

\item{CI_level}{The posterior probability to be contained in the credible
interval.}

\item{seed}{Always set your seed!}

\item{mc_error}{This is the error in probability from the posterior CDF
evaluated at the ROPE bounds. Note that if it is estimated that these
probabilities are between 0.11 and 0.89, the more relaxed value of 0.01 is
used.}
}
\description{
Test of independence for 2-way contingency tables
}
\details{
For a 2-way contingency table with R rows and C columns, evaluate
the probability that
\itemize{
\item the joint probabilities \eqn{p_{ij}} are all
within the ROPE of \eqn{p_{i\cdot}\times p_{\cdot j}} for \code{sampling_design = "multinomial"}
\item the probabilities \eqn{p_{j|i}} are all
within the ROPE of \eqn{p_{\cdot j}} if \code{sampling_design = "fixed rows"} or
\code{"fixed columns"}
}
}
\references{
Gunel, Erdogan &  Dickey, James (1974). Bayes factors for independence in contingency tables, Biometrika, 61(3), Pages 545–557, https://doi.org/10.1093/biomet/61.3.545

Kass, R. E., & Raftery, A. E. (1995). Bayes Factors. Journal of the American Statistical Association, 90(430), 773–795.
}
