% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Extract coefficients from a linreg object}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}.}

\item{...}{Additional arguments.}
}
\value{
A named numeric vector containing the coefficients.
}
\description{
Returns the coefficients estimated by the linear regression model.
}
