% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegTruncClass.R
\name{coef.linregtrunc}
\alias{coef.linregtrunc}
\title{Provide the Coefficient (Parameter Estimates) of the Model}
\usage{
\method{coef}{linregtrunc}(object)
}
\arguments{
\item{object}{an instance of the S3 class linregtrunc}
}
\value{
a named vector
}
\description{
Provide the Coefficient (Parameter Estimates) of the Model
}
