% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.lm_b}
\alias{coef.lm_b}
\title{Coefficient extraction for lm_b or aov_b object}
\usage{
\method{coef}{lm_b}(object)
}
\arguments{
\item{object}{lm_b or aov_b object}
}
\description{
Coefficient extraction for lm_b or aov_b object
}
