% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{Coefficients for ridgereg objects}
\usage{
\method{coef}{ridgereg}(object, ...)
}
\arguments{
\item{object}{A ridgereg object}

\item{...}{Additional arguments}
}
\description{
Coefficients for ridgereg objects
}
