% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegTrunc.R
\name{createDataSet}
\alias{createDataSet}
\title{Create a Data Structure}
\usage{
createDataSet(formula, data)
}
\arguments{
\item{formula}{a formula (e.g. "y ~ x")}

\item{data}{a data.frame object}
}
\description{
Create a data structure on the Java end that will be later used with
the GLM or SIMEX methods.
}
