% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegTrunc.R
\docType{data}
\name{datasetSingleObs}
\alias{datasetSingleObs}
\title{A data.frame object for an example of regression with
residual error terms following a truncated normal distribution.}
\format{
An object of class \code{data.frame} with 5641 rows and 8 columns.
}
\usage{
data(datasetSingleObs)
}
\description{
A data.frame object for an example of regression with
residual error terms following a truncated normal distribution.
}
\examples{
data(datasetSingleObs)
}
\keyword{datasets}
