% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{ext_print}
\alias{ext_print}
\title{Function to use print in class}
\usage{
ext_print(to_print)
}
\arguments{
\item{to_print}{String that should be printed}
}
\description{
Use print() function inside the class methods
}
