% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_invgamma_parms.R
\name{find_invgamma_parms}
\alias{find_invgamma_parms}
\title{Find parameters for Inverse gamma prior based on prior mean and one quantile}
\usage{
find_invgamma_parms(mean, quantile, left_tail_prob, plot_results = TRUE)
}
\arguments{
\item{mean}{numeric greater than 0 giving the prior mean}

\item{quantile}{numeric greater than 0 giving the quantile lying at left_tail_prob}

\item{left_tail_prob}{numeric between 0 and 1 giving the prior probability of theta
being less than or equal to quantile}
}
\description{
Find parameters for Inverse gamma prior based on prior mean and one quantile
}
\examples{
find_invgamma_parms(10 / 4,4.1,0.9)

}
