% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{fitted.linreg}
\alias{fitted.linreg}
\title{Extract fitted values from linreg objects}
\usage{
\method{fitted}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class "linreg"}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of fitted values
}
\description{
Extract fitted values from linreg objects
}
