% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegTruncClass.R
\name{fitted.linregtrunc}
\alias{fitted.linregtrunc}
\title{Provide the Model Predictions}
\usage{
\method{fitted}{linregtrunc}(object)
}
\arguments{
\item{object}{an instance of the S3 class linregtrunc}
}
\value{
a vector
}
\description{
Provide the Model Predictions
}
