% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{fitted.ridgereg}
\alias{fitted.ridgereg}
\title{Fitted values for ridgereg objects}
\usage{
\method{fitted}{ridgereg}(object, ...)
}
\arguments{
\item{object}{A ridgereg object}

\item{...}{Additional arguments}
}
\description{
Fitted values for ridgereg objects
}
