% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegTruncClass.R
\name{fittedOriginalScale}
\alias{fittedOriginalScale}
\title{Provide the Model Predictions on the Original Scale}
\usage{
fittedOriginalScale(object)
}
\arguments{
\item{object}{an instance of the S3 class linregtrunc}
}
\value{
a vector
}
\description{
If a log transformation was used, the model predictions are
back transformed to the original scale.
}
