% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{fix_numeric}
\alias{fix_numeric}
\title{Make sure all columns containing numbers are numeric}
\usage{
fix_numeric(data)
}
\arguments{
\item{data}{dataframe containing data read from the file}
}
\value{
dataframe after marking numeric columns correctly
}
\description{
Fix numeric columns marked as non-numeric while reading data
}
