% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_CIs_from_BMA.R
\name{get_est_and_ci}
\alias{get_est_and_ci}
\title{Estimates and CI from BMA}
\usage{
get_est_and_ci(
  object,
  data,
  y_name,
  covar_names,
  CI_level = 0.95,
  mc_draws = 50000
)
}
\arguments{
\item{object}{An object returned from BMS::bms().  Make sure you return
a reasonably large number of models (nmodel)}

\item{data}{Data used in linear regression model}

\item{y_name}{character for response variable.}

\item{covar_names}{character vector for covariates in fullest model}

\item{CI_level}{Level for credible interval}
}
\description{
Estimates and CI from BMA
}
