% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heteroscedasticity_test.R
\name{heteroscedasticity_test}
\alias{heteroscedasticity_test}
\alias{var_test_b}
\title{Test for heteroscedasticity in AOV models}
\usage{
heteroscedasticity_test(hetero_model, homo_model)
}
\arguments{
\item{hetero_model}{aov_b object where the heteroscedastic argument
has been set to TRUE (the default)}

\item{homo_model}{aov_b object where the heteroscedastic argument
has been set to FALSE}
}
\description{
Use Chib's method to compute the Bayes factor to test for
heteroscedasticity in analysis of variance models.
}
\references{
Kass, R. E., & Raftery, A. E. (1995). Bayes Factors. Journal of the American Statistical Association, 90(430), 773–795.
}
