% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC.R
\name{IC}
\alias{IC}
\alias{DIC}
\alias{AIC}
\alias{BIC}
\alias{BIC.lm_b}
\alias{AIC.lm_b}
\alias{DIC.lm_b}
\alias{BIC.aov_b}
\alias{AIC.aov_b}
\alias{DIC.aov_b}
\title{Compute AIC, BIC, or DIC for aov_b or lm_b objects.  (Lower is better.)}
\usage{
DIC(object)

\method{BIC}{lm_b}(object)

\method{AIC}{lm_b}(object)

\method{DIC}{lm_b}(object)

\method{BIC}{aov_b}(object)

\method{AIC}{aov_b}(object)

\method{DIC}{aov_b}(object)
}
\description{
Compute AIC, BIC, or DIC for aov_b or lm_b objects.  (Lower is better.)
}
