% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{length_df}
\alias{length_df}
\title{Get Dataframe Length}
\usage{
length_df(df)
}
\arguments{
\item{df}{this function takes a dataframe with 'y' column}
}
\value{
length_of_dependent_vector
}
\description{
Get Dataframe Length
}
