\name{linreg-package}
\alias{linreg-package}
\alias{linreg}
\docType{package}
\title{
  \packageTitle{linreg}
}
\description{
  \packageDescription{linreg}
}
\details{
  
  The DESCRIPTION file:
    \packageDESCRIPTION{linreg}
  \packageIndices{LinReg}
  An overview of how to use the package, including the 
  most important functions
}
\author{
  \packageAuthor{linreg}
  
  Maintainer: \packageMaintainer{linreg}
}
\references{
  Literature or other references for background information 
}

\keyword{ package }
\seealso{
  Optional links to other man pages, e.g.
  \code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
  linreg(Sepal.Length ˜ Sepal.Width, iris)
}

