% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_class.R
\docType{class}
\name{linreg_class-class}
\alias{linreg_class}
\alias{linreg_class-class}
\title{the linreg class}
\description{
the linreg class
}
\examples{
linreg(Sepal.Length ~ Sepal.Width, iris)
}
\author{
Yumeng Li, Mattias Karlsson, Ashraf Sarhan
}
