% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linreg.R
\name{linreg_f}
\alias{linreg_f}
\title{Linear Regression Function}
\usage{
linreg_f(formula, data)
}
\arguments{
\item{formula}{formula (y ~ x)}

\item{data}{dataframe}
}
\value{
An object of RC class linreg
}
\description{
This function takes two arguments- formula and data,
    and returns an object of class linreg which builds a linear
    regression model
}
\examples{
linreg_f(formula = Petal.Length ~ Species, data = iris)
}
