% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegTrunc.R
\name{LinRegTrunc}
\alias{LinRegTrunc}
\title{Correct Inference for Generalized Linear Models with Measurement Errors
using the SIMEX Method}
\usage{
LinRegTrunc(formula, data, truncation, isLogTransformed = T, constant = 1)
}
\arguments{
\item{formula}{a formula (e.g. "y ~ x")}

\item{data}{a data.frame object}

\item{truncation}{a numeric that sets the lower bound of the truncated distribution}

\item{isLogTransformed}{a logical, it is assumed the response variable has been
log transformed}

\item{constant}{a constant that has been added to the response variable before
the log transformation. It is assumed that this constant is 1.}
}
\value{
an instance of the S3 SIMEXResult class
}
\description{
First, it create a data structure on the Java end that will be later used with
the SIMEX method. Secondly, it fits a naive model. Thirdly, it implements the
SIMEX method.
}
