% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_b.R
\name{lmer_b}
\alias{lmer_b}
\alias{glmer_b}
\title{Bayesian Mixed Effects Regression}
\usage{
lmer_b(
  ...,
  fixef_df = 5,
  fixef_mean = 1,
  fixef_scale = c(10^2, 2.5^2),
  fixef_common_scale = TRUE,
  .use_t_prior = TRUE
)
}
\arguments{
\item{formula}{a two-sided linear formula object describing both the
fixed-effects and random-effects part of the model, with the response on
the left of a \code{~} operator and the terms, separated by \code{+}
operators, on the right. Random-effects terms are distinguished by
vertical bars (\code{|}) separating expressions for design matrices
from grouping factors. Two vertical bars (\code{||}) can be used to
specify multiple uncorrelated random effects for the same grouping
variable. (Because of the way it is implemented, the \code{||}-syntax
\emph{works only for design matrices containing numeric (continuous) predictors};
to fit models with independent categorical effects, see \link[lme4]{dummy}
or the \code{lmer_alt} function from the \link[afex]{\code{afex-package}} package.)}

\item{data}{an optional data frame containing the variables named in
formula. By default the variables are taken from the environment from which
\code{lmer_b} is called. While data is optional, the package authors
\emph{strongly} recommend its use, especially when later applying methods such
as \code{update} and \code{drop1} to the fitted model (\emph{such methods are
not guaranteed to work properly if data is omitted}). If \code{data} is
omitted, variables will be taken from the environment of \code{formula}
(if specified as a formula) or from the parent frame (if specified as a
character vector).}

\item{family}{A description of the error distribution and link function
to be used in the model. See ?glm for more information.}

\item{fixef.prior}{a BLME prior of family \code{normal}, \code{t}, or
\code{horseshoe}. Imposes a prior over the fixed
effects/modeled coefficients. Default is \code{normal} centered at zero
with variance equal to \eqn{2.5^2\frac{\text{max}(1,s_y^2)}{s_j^2}},
where \eqn{s_y} is the standard deviation of the outcome/response
variable, and \eqn{s_j} is the standard deviation of the \eqn{j^{th}}
covariate.  The intercept, however, differs in that it is centered at
\eqn{\bar{y}}, and has prior variance equal to \eqn{2.5^2s_y^2}.
See \code{vignette("Regression-methods","bayesics")}
for more on the motivation for this prior.}

\item{resid.prior}{a BLME prior of family \code{gamma}, \code{invamma},
or \code{point}. Imposes a prior over the noise/residual
variance, also known as common scale parameter or the conditional variance
given the random effects. Default is a flat gamma distribution
(\eqn{Gamma(0.001,0.001)}).}

\item{cov.prior}{a BLME prior or list of priors with allowable
distributions: \code{wishart}, \code{invwishart}, \code{gamma}, or
\code{invgamma}. Imposes a prior over the covariance of the random
effects/modeled coefficients. Default is \code{wishart}.}

\item{ROPE}{vector of positive values giving ROPE boundaries for each regression
coefficient.  Optionally, you can not include a ROPE boundary for the intercept.
If missing, defaults go to those suggested by Kruchke (2018).}

\item{control}{a list (of correct class, resulting from \link[lme4]{lmerControl}()
or \link[lme4]{glmerControl}() respectively) containing control parameters,
including the nonlinear optimizer to be used and parameters to be passed
through to the nonlinear optimizer, see the \code{*lmerControl}
documentation in the \link[lme4]{\code{lme4-package}} package for details.}

\item{start}{a named list of starting values for the parameters in the
model. For \code{lmer_b} this can be a numeric vector or a list with one
component named \code{"theta"}.  For \code{glmer_b}, a numeric \code{start}
argument will be used as the starting value of \code{theta}. If start is a
list, the \code{theta} element (a numeric vector) is used as the starting
value for the first optimization step (default=1 for diagonal elements and
0 for off-diagonal elements of the lower Cholesky factor); the fitted value
of theta from the first step, plus \code{start[["fixef"]]}, are used as
starting values for the second optimization step. If \code{start} has both
\code{fixef} and \code{theta} elements, the first optimization step is
skipped. For more details or finer control of optimization, see
\link[lme4]{modular}.}

\item{verbose}{integer scalar. If \code{> 0} verbose output is generated
during the optimization of the parameter estimates. If \code{> 1} verbose
output is generated during the individual penalized iteratively reweighted
least squares (PIRLS) steps.}

\item{nAGQ}{integer scalar - the number of points per axis for evaluating
the adaptive Gauss-Hermite approximation to the log-likelihood. Defaults to
1, corresponding to the Laplace approximation. Values greater than 1
produce greater accuracy in the evaluation of the log-likelihood at the
expense of speed. A value of zero uses a faster but less exact form of
parameter estimation for GLMMs by optimizing the random effects and the
fixed-effects coefficients in the penalized iteratively reweighted least
squares step. (See Details of \link[lme4]{glmer}.)}

\item{subset}{an optional expression indicating the subset of the rows of
data that should be used in the fit. This can be a logical vector, or a
numeric vector indicating which observation numbers are to be included, or
a character vector of the row names to be included. All observations are
included by default.}

\item{weights}{an optional vector of \code{‘prior weights’} to be used in
the fitting process. Should be NULL or a numeric vector.}

\item{na.action}{a function that indicates what should happen when the
data contain NAs. The default action (\code{na.omit}, inherited from the
‘factory fresh’ value of \code{getOption("na.action")}) strips any
observations with any missing values in any variables.}

\item{offset}{this can be used to specify an a priori known component to be
included in the linear predictor during fitting. This should be \code{NULL}
or a numeric vector of length equal to the number of cases. One or more
offset terms can be included in the formula instead or as well,
and if more than one is specified their sum is used.}

\item{contrasts}{an optional list. See the \code{contrasts.arg} of
\link[stats]{model.matrix.default}.}

\item{mustart}{optional starting values on the scale of the conditional
mean, as in \link[stats]{glm}; see there for details.}

\item{etastart}{optional starting values on the scale of the
unbounded predictor as in \link[stats]{glm}; see there for details.}

\item{devFunOnly}{logical - return only the deviance evaluation function.
Note that because the deviance function operates on variables stored in
its environment, it may not return \emph{exactly} the same values on subsequent
calls (but the results should always be within machine tolerance).}
}
\description{
Maximum  a posteriori estimation for linear and generalized linear
mixed-effects models in a Bayesian setting.  \strong{This is a wrapper for
\link[blme]{blmer} and \link[blme]{bglmer}, which in turn is built off of
\link[lme4]{lmer} and \link[lme4]{glmer}.}
}
\details{
The bulk of the usage for \code{blmer} and \code{bglmer} closely
follows the functions \code{\link[lme4]{lmer}} and
\code{\link[lme4]{glmer}}. Those help pages provide a good overview of
fitting linear and generalized linear mixed models. The primary
distinction is that \code{blmer} and \code{bglmer} allow the user to
do Bayesian inference or penalized maximum likelihood, with priors imposed on the different
model components. For the specifics of any distribution listed below,
see the \link[=bmerDist-class]{distributions} page.

\strong{Covariance Prior}

The \code{cov.prior} argument applies a prior over the
covariance matrix of the random effects/modeled coefficients.
As there is one covariance matrix for every named grouping factor -
that is every element that appears to the right of a vertical bar
("|") in the model formula - it is possible to apply as many
different priors as there are said factors.

The general formats of an argument to \code{blmer} or \code{bglmer}
for such a prior are of the form:

\itemize{
\item \code{cov.prior = factor.name ~ covariance.distribution(option1 = value1, \dots)}
\item \code{cov.prior = list(fc.nm ~ dist1, fc.nm ~ dist2, ..., default.distribution)}
}

If the \dQuote{\code{factor.name ~}} construct is omitted, the prior
is interpretted as a default and applied to all factors that
lack specific priors of their own. Options are not required,
but permit fine-tuning of the model.

Supported distributions are \code{gamma}, \code{invgamma}, \code{wishart},
\code{invwishart}, \code{NULL}, and \code{custom}.

The \code{common.scale} option, a logical, determines whether or
not the prior applies to in the absolute-real world
sense (value = \code{FALSE}), or if the prior is applied to the random effect
covariance divided by the estimated residual variance (\code{TRUE}). As a practical matter,
when false computation can be slower as the profiled common scale may
no longer have a closed-form solution. As such, the default for all
cases is \code{TRUE}.

Other options are specified along with the specific distributions and
defaults are explained in the blme \link[=bmerDist-class]{distributions} page.

\strong{Fixed Effects Prior}

Priors on the fixed effects, or unmodeled coefficients, are specified
in a fashion similar to that of covariance priors. The general format is

\itemize{\item \code{fixef.prior = multivariate.distribution(options1 = value1, \dots)}}

At present, the implemented multivariate distributions are \code{normal}, \code{t},
\code{horseshoe}, and \code{NULL}. \code{t} and \code{horseshoe} priors cannot be used
when \code{REML} is \code{TRUE}, as that integral does not have a closed form solution.

\strong{Residual Variance Prior}
The general format for a residual variance prior is the same as for a fixed
effect prior. The supported distributions are \code{point}, \code{gamma},
\code{invgamma}.
}
\references{
Douglas Bates, Martin Maechler, Ben Bolker, Steve Walker (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1-48. doi:10.18637/jss.v067.i01.

Chung Y, Rabe-Hesketh S, Dorie V, Gelman A, Liu J (2013). “A nondegenerate penalized likelihood estimator for variance parameters in multilevel models.” \emph{Psychometrika}, \emph{78}(4), 685-709. doi:10.1007/s11336-013-9328-2 \url{https://doi.org/10.1007/s11336-013-9328-2}.
}
