\name{luckylab2.0-package}
\alias{luckylab2.0-package}
\alias{luckylab2.0}
\docType{package}
\title{
\packageTitle{luckylab2.0}
}
\description{
\packageDescription{luckylab2.0}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{luckylab2.0}
\packageIndices{luckylab2.0}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{luckylab2.0}

Maintainer: \packageMaintainer{luckylab2.0}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
