% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{mean_for_x}
\alias{mean_for_x}
\title{Calculate Mean of X}
\usage{
mean_for_x(df)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'x' column}
}
\value{
mean_of_independent_vector
}
\description{
Calculate Mean of X
}
