% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{mean_for_y}
\alias{mean_for_y}
\title{Calculate Mean of Y}
\usage{
mean_for_y(df)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'y' column}
}
\value{
mean_of_dependent_vector
}
\description{
Calculate Mean of Y
}
