% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multreg.R
\name{multreg}
\alias{multreg}
\title{Regression}
\usage{
multreg(Q, R, y)
}
\arguments{
\item{Q}{is matrix formed via QR decomposition}

\item{R}{is a triangular matrix formed via QR decomposition}

\item{y}{is the variable that is to be regressed on}
}
\value{
list of the calculated coefficients
}
\description{
Regression
}
\examples{

}
