% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_lm_b.R
\name{np_lm_b}
\alias{np_lm_b}
\title{Non-parametric linear models}
\usage{
np_lm_b(formula, data, family, trials, n_draws, CI_level = 0.95, seed = 123)
}
\arguments{
\item{formula}{A formula specifying the model.}

\item{data}{A data frame in which the variables specified in the formula
will be found. If missing, the variables are searched for in the standard way.}

\item{family}{A description of the error distribution and link function
to be used in the model. See ?glm for more information.}

\item{trials}{Integer vector giving the number of trials for each
observation if family = binomial().}

\item{n_draws}{integer.  Number of posterior draws to obtain.}

\item{CI_level}{numeric. Credible interval level.}

\item{seed}{integer.  Fed into either set.seed() as the argument future.seed.}
}
\value{
np_lm_b() returns an object of class "np_lm_b", which behaves as
a list with the following elements:
\itemize{
\item summary - a tibble giving results for regression coefficients.
}
}
\description{
np_lm_b uses general Bayesian inference with loss-likelihood bootstrap.
This is, as implemented here, a Bayesian non-parametric linear models
inferential engine, using the generalized least squares loss function.
}
\details{
The GLS loss function is \eqn{\sum_i (y_i - \mu_i)^2/V(\mu_i)}, where
\eqn{g(\mu_i) = X_i'\beta + \text{offset}} for some appropriate link
function. Applicable data types are continuous (use family =
gaussian()), count (use family = poisson()), or binomial
(use family = binomial()).
}
