% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Diagnostic plots for a linreg object}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{linreg}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{NULL} and displays two diagnostic plots.
}
\description{
Creates a Residuals vs Fitted plot and a Scale-Location plot.
}
