% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.lm_b}
\title{Print linreg objects.}
\usage{
\method{plot}{lm_b}(
  x,
  type = c("diagnostics", "pdp", "ci band", "pi band")[1],
  variable,
  exemplar_covariates,
  combine_pi_ci = TRUE,
  variable_seq_length = 100,
  return_as_list = FALSE
)
}
\arguments{
\item{x}{A linreg object}

\item{type}{character. Select any of "diagnostics" ("dx" is also allowed),
"pdp" (partial dependence plot), "ci band", and/or "pi band".  NOTE: the CI
and PI bands only work for numeric variables.}

\item{variable}{character. If type = "pdp" , which variable should be plotted?}

\item{exemplar_covariates}{data.frame or tibble with exactly one row.
Used to fix other covariates while varying the variable of interest for the plot.}

\item{combine_pi_ci}{logical. If type includes both "ci band" and "pi band",
should the credible band be superimposed on the prediction band or
plotted separately?}

\item{variable_seq_length}{integer. Number of points used to draw pdp.}

\item{return_as_list}{logical.  If TRUE, a list of ggplots will be returned,
rather than a single plot produced by the patchwork package.}
}
\description{
Print linreg objects.
}
