% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_test_b.R
\name{poisson_test_b}
\alias{poisson_test_b}
\title{Poisson tests}
\usage{
poisson_test_b(
  x,
  offset,
  r,
  ROPE,
  prior = "jeffreys",
  prior_shape_rate,
  CI_level = 0.95,
  plot = TRUE,
  seed = 1,
  mc_relative_error = 0.01
)
}
\arguments{
\item{x}{Number of events.  A vector of length one or two.}

\item{offset}{Time, area, etc. measured in the Poisson process.  NOTE: Do not
take the log!}

\item{r}{optional.  If provided and inference is being made for
a single population, \code{poisson_test_b} will return the posterior
probability that the population rate is less than this value.}

\item{ROPE}{ROPE for rate ratio if inference is being made for two populations.
Provide either a single value or a vector of length two.  If the former,
the ROPE will be taken as (1/ROPE),ROPE).  If the latter, these will be
the bounds of the ROPE.}

\item{prior}{Either "jeffreys" (Gamma(1/2,0)) or "flat" (Gamma(0.001,0.001)).
This is ignored if prior_shape_rate is provided.}

\item{prior_shape_rate}{Vector of length two, giving the shape and rate parameters
for the gamma distribution that will act as the prior on the population
rates.}

\item{CI_level}{The posterior probability to be contained in the
credible intervals.}

\item{plot}{logical.  Should a plot be shown?}

\item{seed}{Always set your seed!  (Unused for a single population rate)}

\item{mc_relative_error}{The relative monte carlo error of the quantiles of the CIs.
(Ignored for a single population rate)}
}
\description{
Make inference on one or two populations of Poisosn distributed count data
}
\details{
The likelihood is
\deqn{
 y \sim Poi(\lambda t),
}
where \eqn{\lambda} is the rate, and \eqn{t} is the time or area observed
and is given by the argument \code{offset}.

The prior is given by
\deqn{
 \lambda \sim \Gamma(a,b),
}
where \eqn{a} and \eqn{b} are given by the argument \code{prior_shape_rate}.
If \code{prior_shape_rate} is missing and \code{prior = "jeffreys"},
then a Jeffrey's prior will be used, i.e., \eqn{\Gamma(0.5,0)} (improper),
while if \code{prior = "flat"}, \eqn{\Gamma(0.001,0.001)} will be used.
}
