% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Shorthand predict method for linreg objects}
\usage{
\method{pred}{linreg}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class "linreg"}

\item{newdata}{A data frame with new data for prediction}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of predicted values
}
\description{
Shorthand predict method for linreg objects
}
