% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\alias{pred.linreg}
\title{Extract predicted values}
\usage{
pred(object, ...)

\method{pred}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector containing the predicted values.
}
\description{
Returns the fitted values estimated by a linreg model.
}
