% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bma.R
\name{predict.bma}
\alias{predict.bma}
\title{Predict method for bma model fits}
\usage{
\method{predict}{bma}(object, newdata, CI_level = 0.95, future.seed = 1)
}
\arguments{
\item{object}{Object of class bma}

\item{newdata}{An optional data.frame in which to look for variables with which
to predict.}

\item{CI_level}{numeric. Credible interval level.}
}
\value{
tibble with estimate, prediction intervals, and credible intervals
for the mean.
}
\description{
Predict method for bma model fits
}
