% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.glm_b.R
\name{predict.glm_b}
\alias{predict.glm_b}
\title{Predict method for lm_b model fits}
\usage{
\method{predict}{glm_b}(object, newdata, trials, CI_level = 0.95, PI_level = 0.95, seed = 1)
}
\arguments{
\item{object}{Object of class lm_b}

\item{newdata}{An optional data.frame in which to look for variables with which
to predict.}

\item{trials}{Integer vector giving the number of trials for each
observation if family = binomial().}

\item{CI_level}{Posterior probability covered by credible interval}

\item{PI_level}{Posterior probability covered by prediction interval}

\item{seed}{integer.  Always set your seed!!!}
}
\value{
tibble with estimate (posterior mean), prediction intervals, and credible intervals
for the mean.
}
\description{
Predict method for lm_b model fits
}
