% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{predict.linreg}
\alias{predict.linreg}
\title{Predict method for linreg objects}
\usage{
\method{predict}{linreg}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class "linreg"}

\item{newdata}{A data frame with new data for prediction}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of predicted values
}
\description{
Predict method for linreg objects
}
