% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.lm_b_bma.R
\name{predict.lm_b_bma}
\alias{predict.lm_b_bma}
\title{Predict method for bma model fits}
\usage{
\method{predict}{lm_b_bma}(object, newdata, CI_level = 0.95, PI_level = 0.95, seed = 1)
}
\arguments{
\item{object}{Object of class bma}

\item{newdata}{An optional data.frame in which to look for variables with which
to predict.}

\item{CI_level}{Posterior probability covered by credible interval}

\item{PI_level}{Posterior probability covered by prediction interval}

\item{seed}{integer. Always set your seed!!!}
}
\value{
list.
\itemize{
\item newdata tibble with estimate, prediction intervals, and credible intervals
for the mean.
\item posterior_draws
\itemize{
\item mean_of_ynew draws of \eqn{E(y)}, marginalizing out the model
\item posterior draws of ynew
}
}
}
\description{
Predict method for bma model fits
}
