% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.np_lm_b.R
\name{predict.np_lm_b}
\alias{predict.np_lm_b}
\title{Predict method for lm_b model fits}
\usage{
\method{predict}{np_lm_b}(object, newdata, trials, CI_level = 0.95, n_draws = 0)
}
\arguments{
\item{object}{Object of class lm_b}

\item{newdata}{An optional data.frame in which to look for variables with which
to predict.}

\item{CI_level}{numeric. Credible interval level.}
}
\value{
tibble with estimate, prediction intervals, and credible intervals
for the mean.
}
\description{
Predict method for lm_b model fits
}
