% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{Predict method for ridgereg objects}
\usage{
\method{predict}{ridgereg}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A ridgereg object}

\item{newdata}{New data frame for prediction}

\item{...}{Additional arguments}
}
\description{
Predict method for ridgereg objects
}
