% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.linreg}
\alias{print.linreg}
\title{Print a linreg object}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{linreg}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns the original \code{linreg} object.
}
\description{
Prints the model call and its estimated coefficients.
}
