% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.aov_b}
\alias{print.lm_b}
\title{Print aov_b or lm_b objects.}
\usage{
\method{print}{aov_b}(object)

\method{print}{lm_b}(object)
}
\description{
Print aov_b or lm_b objects.
}
