% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{Print for ridgereg class}
\usage{
\method{print}{ridgereg}(x, ...)
}
\arguments{
\item{x}{a ridgereg class object}

\item{...}{optional arguments to pass to generic}
}
\value{
A print for the ridgereg class
}
\description{
Print for ridgereg class
}
\examples{
data(mtcars)
model <- ridgereg(mpg~wt+cyl, mtcars, lambda = 0, QR = FALSE)
print(model)
}
