% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR.R
\name{QR}
\alias{QR}
\title{QR method}
\usage{
QR(X)
}
\arguments{
\item{X}{is a matrix for which the Q and the R matrix will be calculated from}
}
\value{
The Q and R matrix where R is triangular
}
\description{
QR method
}
\examples{

}
