% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{read_data}
\alias{read_data}
\title{Load dataset from files of different formats}
\usage{
read_data(loc)
}
\arguments{
\item{loc}{file location}
}
\value{
dataframe containing data read from the file
}
\description{
Read data from different file formats into a dataframe. Supported formats are csv, txt, fwf, tsv, psv, xls, xlsx, json, sav, sas7bdat, rds, dta.
}
