% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{resid}
\alias{resid}
\title{Residuals}
\usage{
resid(x)
}
\arguments{
\item{x}{A class object of Linear Regression}
}
\value{
the vector of residuals e
}
\description{
\code{resid.linreg} return the vector of residuals e
}
\details{
This function returns the vector of residuals e stored in 
the object linreg of S3 class.
}
