% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Extract residuals from a linreg object}
\usage{
\method{residuals}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector containing the residuals.
}
\description{
Returns the residuals estimated by the linear regression model.
}
