% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{residuals.ridgereg}
\alias{residuals.ridgereg}
\title{Residuals for ridgereg objects}
\usage{
\method{residuals}{ridgereg}(object, ...)
}
\arguments{
\item{object}{A ridgereg object}

\item{...}{Additional arguments}
}
\description{
Residuals for ridgereg objects
}
