% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Title}
\description{
Title
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{B_ridge}}{vector.}

\item{\code{y_h}}{matrix.}

\item{\code{X}}{matrix.}

\item{\code{data_name}}{character.}

\item{\code{lamda}}{numeric.}
}}


