% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ridgereg.R
\name{ridgereg_f}
\alias{ridgereg_f}
\title{Ridge Regression Function}
\usage{
ridgereg_f(formula, data)
}
\arguments{
\item{formula}{formula (y ~ x)}

\item{data}{dataframe}
}
\value{
An object of RC class ridgereg
}
\description{
This function takes two arguments- formula and data,
    and returns an object of class ridgereg which builds a ridge
    regression model
}
\examples{
ridgereg_f(formula = Petal.Length ~ Species, data = iris)
}
