% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg_fit}
\alias{ridgereg_fit}
\title{Fit a Ridge Regression Model}
\usage{
ridgereg_fit(formula, data, lambda)
}
\arguments{
\item{formula}{A formula object specifying the model to be fitted.}

\item{data}{A data frame containing the dataset for fitting.}

\item{lambda}{A numeric value specifying the ridge penalty parameter.}
}
\value{
An instance of the `ridgereg` class containing the fitted model.
}
\description{
Fit a Ridge Regression Model
}
\examples{
data(iris)
mod <- ridgereg_fit(Petal.Length ~ Species, data = iris, lambda = 0.1)
mod$print()
mod$predict(newdata = iris)
mod$coef()
}
\seealso{
\code{\link[MASS]{lm.ridge}} for a comparable ridge regression implementation in the MASS package.
}
