% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDratio.R
\name{SDratio}
\alias{SDratio}
\title{Bayes factors}
\usage{
SDratio(object)
}
\arguments{
\item{object}{lm_b object}
}
\description{
Bayes factors for Bayesian regression objects using the Savage-Dickey ratio
}
\details{
Bayes factors are given in terms of favoring the two-tailed alternative hypothesis
vs. the null hypothesis that the regression coefficient equals zero.  Currently only
implemented for lm_b objects.  Use bridgesampling package for aov_b.
}
