% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{significance}
\alias{significance}
\title{Significance
Helper function to encode significance based on a given p_value}
\usage{
significance(p_value)
}
\arguments{
\item{p_value}{The given p_value}
}
\description{
Significance
Helper function to encode significance based on a given p_value
}
