% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegTruncClass.R
\name{summary.linregtrunc}
\alias{summary.linregtrunc}
\title{Provide a Summary of the fit}
\usage{
\method{summary}{linregtrunc}(object, ...)
}
\arguments{
\item{object}{an instance of the S3 class linregtrunc}

\item{...}{absolutely useless, just to keep the signature consistent for inheritance}
}
\description{
Provide a Summary of the fit
}
