% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.lm_b}
\alias{summary.aov_b}
\alias{summary.np_lm_b}
\title{Summary functions for linreg objects}
\usage{
\method{summary}{lm_b}(object, CI_level = 0.95)

\method{summary}{aov_b}(object, CI_level = 0.95)

\method{summary}{np_lm_b}(object)
}
\arguments{
\item{object}{linreg object}
}
\description{
Summary functions for linreg objects
}
