% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{summary.ridgereg}
\alias{summary.ridgereg}
\title{Summary method for ridgereg objects}
\usage{
\method{summary}{ridgereg}(object, ...)
}
\arguments{
\item{object}{A ridgereg object}

\item{...}{Additional arguments}
}
\description{
Summary method for ridgereg objects
}
