% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit_b.R
\name{survfit_b}
\alias{survfit_b}
\title{Create survival curves}
\usage{
survfit_b(formula, data, prior_shape, prior_rate, max_n_time_bins, n_time_bins)
}
\description{
Use the semi-parametric piecewise exponential survival model
to fit a survival curve
}
\references{
Qing Y, Thall PF, Yuan Y. A Bayesian piecewise exponential phase II design for monitoring a time-to-event endpoint. Pharm Stat. 2023 Jan;22(1):34-44. doi: 10.1002/pst.2256. Epub 2022
}
