% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{Total_sum_x}
\alias{Total_sum_x}
\title{Calculate total sum for x}
\usage{
Total_sum_x(df)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'x' column}
}
\value{
total_sum_of_independent_vector
}
\description{
Calculate total sum for x
}
