% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr2.R
\name{Total_sum_y}
\alias{Total_sum_y}
\title{Calculate total sum for y}
\usage{
Total_sum_y(df)
}
\arguments{
\item{df}{this funtion takes a dataframe with 'y' column}
}
\value{
total_sum_of_dependent_vector
}
\description{
Calculate total sum for y
}
