% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegClass.R
\name{vcov.LinReg}
\alias{vcov.LinReg}
\title{Provide the Estimated Variance-Covariance of the Model Coefficients}
\usage{
\method{vcov}{LinReg}(object)
}
\arguments{
\item{object}{an instance of the S3 class LinReg}
}
\value{
a matrix
}
\description{
Provide the Estimated Variance-Covariance of the Model Coefficients
}
