% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegTruncClass.R
\name{vcov.linregtrunc}
\alias{vcov.linregtrunc}
\title{Provide the Estimated Variance-Covaraince of the Model Coefficients}
\usage{
\method{vcov}{linregtrunc}(object)
}
\arguments{
\item{object}{an instance of the S3 class linregtrunc}
}
\value{
a matrix
}
\description{
Provide the Estimated Variance-Covaraince of the Model Coefficients
}
