% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_test_b.R
\name{wilcoxon_test_b}
\alias{wilcoxon_test_b}
\title{Bayesian Wilcoxon Rank Sum and Signed Rank Analyses}
\usage{
wilcoxon_test_b(
  x,
  y,
  paired = FALSE,
  p = 0.5,
  ROPE,
  prior = "centered",
  prior_shapes,
  CI_level = 0.95,
  plot = TRUE,
  seed = 1
)
}
\arguments{
\item{x}{numeric vector of data values. Non-finite (e.g., infinite or
missing) values will be omitted.}

\item{y}{an optional numeric vector of data values: as with x non-finite values will be omitted.}

\item{paired}{if \code{TRUE} and \code{y} is supplied, x-y will be the input of the Bayesian
Wilcoxon signed rank test.}

\item{p}{\itemize{
\item Signed rank: If provided, \code{wilcox_test_b} will return the
posterior probability that the population proportion is less than this value.
\item Rank sum:
}}

\item{ROPE}{\itemize{
\item Signed rank: If a single number, ROPE will be \code{p}\eqn{\pm}\code{ROPE}.
If a vector of length 2, these will serve as the ROPE bounds. Defaults to
\eqn{\pm 0.05}.
\item Rank sum:
}}

\item{prior}{Prior used on the probability that x > y.  Either
"jeffreys" (Beta(1/2,1/2)), "uniform" (Beta(1,1)), or "centered" (Beta(2,2)).
This is ignored if prior_shapes is provided.}

\item{prior_shapes}{Vector of length two, giving the shape parameters for the
beta distribution that will act as the prior on the population proportions.}

\item{CI_level}{The posterior probability to be contained in the credible interval.}

\item{plot}{logical.  Should a plot be shown?}

\item{seed}{Always set your seed! (Unused for \eqn{\geq} 20 observations.)}
}
\description{
Bayesian Wilcoxon Rank Sum and Signed Rank Analyses
}
\details{
\strong{Bayesian Wilcoxon signed rank analysis}
For a single input vector or paired data, the Bayesian signed rank
analysis will be performed.  The estimand is the proportion of (differenced)
values that are positive.

\strong{Bayesian Wilcoxon rank sum/Mann-Whitney analysis}
For unpaired x and y inputs, the Bayesian rank sum analysis will be performed.
The estimand is \eqn{\Omega_x:=\lim_{n\to\infty} \frac{U_x}{U_x + U_y}}, where
\eqn{U_x} is the number of pairs \eqn{(i,j)} such that \eqn{x_i > y_j}, and
vice versa for \eqn{U_y}.  That is, it is the population proportion of all
untied pairs for which \eqn{x > y}.  Larger values imply that \eqn{x} is
stochastically larger than \eqn{y}.
}
