% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_b.R
\name{aov_b}
\alias{aov_b}
\title{Analysis of Variance using Bayesian methods}
\usage{
aov_b(
  formula,
  data,
  heteroscedastic = TRUE,
  prior_mean_mu = 0,
  prior_mean_nu = 0.001,
  prior_var_shape = 0.001,
  prior_var_rate = 0.001,
  CI_level = 0.95,
  ROPE = 0.1,
  contrasts,
  n_joint_draws = 1000,
  improper = FALSE
)
}
\arguments{
\item{formula}{A formula specifying the model.}

\item{data}{A data frame in which the variables specified in the formula
will be found. If missing, the variables are searched for in the standard way.}

\item{prior_mean_mu}{numeric. Hyperparameter for the a priori mean of the
group means.}

\item{prior_mean_nu}{numeric. Hyperparameter which scales the precision of
the group means.}

\item{prior_var_shape}{numeric. Twice the shape parameter for the inverse gamma prior on
the residual variance(s).  I.e., \eqn{\sigma^2\sim IG}(prior_var_shape/2,prior_var_rate/2).}

\item{prior_var_rate}{numeric. Twice the rate parameter for the inverse gamma prior on
the residual variance(s).  I.e., \eqn{\sigma^2\sim IG}(prior_var_shape/2,prior_var_rate/2).}

\item{CI_level}{numeric. Credible interval level.}

\item{ROPE}{numeric.  Used to compute posterior probability that Cohen's D +/- ROPE}

\item{contrasts}{numeric/matrix. Either vector of length equal to the number of
levels in the grouping variable, or else a matrix where each row is a separate
contrast.}

\item{n_joint_draws}{integer. Number of posterior draws to obtain.}

\item{improper}{logical.  Should we use an improper prior that is proportional
to the inverse of the variance?}

\item{heterscedastic}{logical.  Set to FALSE to assume all groups have
equal variance.}
}
\value{
Object of class "aov_b" with the following elements:
\itemize{
\item summary - data.frame giving the summary of the model parameters
\item pairwise summary - data.frame giving the summary comparing all
factor level means
\item contrasts (if provided) - list with named elements L (the contrasts provided
by the user) and summary.
\item posterior_draws - mcmc object (see coda package) giving the posterior draws
\item formula, data - input by user
\item posterior_parameters -
\itemize{
\item mu_g - the post. means of the group means
\item nu_g - the post. scalars of the precision
\item a_g - the post. shape of the inv. gamma for the group variances
\item b_g - the post. rate of the inv. gamma for the group variances.
}
\item hyperparameters -
\itemize{
\item mu - the prior mean of the group means
\item nu - the prior scalar of the precision
\item a - the prior shape of the inv. gamma for the group variances
\item b - the prior rate of the inv. gamma for the group variances.
}
}
}
\description{
Analysis of Variance using Bayesian methods
}
