% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDratio.R
\name{BF}
\alias{BF}
\alias{SDratio}
\alias{SDratio.glm_b}
\title{Bayes factors for Bayesian regression objects using the Savage-Dickey ratio}
\usage{
SDratio(object)

\method{SDratio}{glm_b}(object)
}
\arguments{
\item{object}{lm_b or glm_b object}
}
\description{
Bayes factors for Bayesian regression objects using the Savage-Dickey ratio
}
\details{
Bayes factors are given in terms of favoring the two-tailed alternative hypothesis
vs. the null hypothesis that the regression coefficient equals zero.  Currently only
implemented for lm_b objects.  Use bridgesampling package for aov_b.

Interpretation is taken from Kass and Raftery.
}
\references{
Kass, R. E., & Raftery, A. E. (1995). Bayes Factors. Journal of the American Statistical Association, 90(430), 773–795.
}
