% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_class2.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Regression coefficients from linreg object}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg object}

\item{...}{Additional arguments}
}
\value{
Numeric vector of regression coefficients
}
\description{
Regression coefficients from linreg object
}
