% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegClass.R
\name{coef.LinReg}
\alias{coef.LinReg}
\title{Provide the Coefficient (Parameter Estimates) of the Model}
\usage{
\method{coef}{LinReg}(object)
}
\arguments{
\item{object}{an instance of the S3 class LinReg}
}
\value{
a named vector
}
\description{
Provide the Coefficient (Parameter Estimates) of the Model
}
