% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Coefficients for linreg class}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{a linreg class object}

\item{...}{optional arguments to pass to generic}
}
\value{
A vector of coefficients for a linreg class object
}
\description{
Coefficients for linreg class
}
\examples{
data(mtcars)
model <- linreg(mpg~wt+cyl, mtcars, QR = FALSE)
coef(model)
}
