% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Extract coefficients from linreg objects}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class "linreg"}

\item{...}{Additional arguments (not used)}
}
\value{
Named vector of coefficients
}
\description{
Extract coefficients from linreg objects
}
