% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{coef}
\alias{coef}
\title{Regression coefficients}
\usage{
coef(x)
}
\arguments{
\item{x}{A class object of Linear Regression}
}
\value{
A coefficients as a named vector
}
\description{
\code{coef.linreg} Returns the coefficients as a named vector.
}
\details{
This function returns the the coefficients as a named vector stored in 
the object linreg of S3 class.
}
