% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef}
\alias{coef}
\alias{coef.lm_b}
\alias{coef.aov_b}
\alias{coef.np_lm_b}
\alias{coef.lm_b_bma}
\title{Coefficient extraction for linreg objects}
\usage{
\method{coef}{lm_b}(object)

\method{coef}{aov_b}(object)

\method{coef}{np_lm_b}(object)

\method{coef}{lm_b_bma}(object)
}
\arguments{
\item{object}{linreg object}
}
\description{
Coefficient extraction for linreg objects
}
