% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{Coefficients for ridgereg class}
\usage{
\method{coef}{ridgereg}(object, ...)
}
\arguments{
\item{object}{a ridge class object}

\item{...}{optional arguments to pass to generic}
}
\value{
A vector of coefficients for a linreg class object
}
\description{
Coefficients for ridgereg class
}
\examples{
data(mtcars)
model <- ridgereg(mpg~wt+cyl, mtcars,lambda = 0, QR = FALSE)
coef(model)
}
