% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegClass.R
\name{fitted.LinReg}
\alias{fitted.LinReg}
\title{Provide the Model Predictions}
\usage{
\method{fitted}{LinReg}(object, type = "transformed", includeVariance = T)
}
\arguments{
\item{object}{an instance of the S3 class LinReg}

\item{type}{either transformed or original (if there was a log transformation)}

\item{includeVariance}{a logical}
}
\value{
a vector (or a matrix if the variance is included)
}
\description{
Provide the Model Predictions
}
