% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_draws.R
\name{get_posterior_draws}
\alias{get_posterior_draws}
\title{Get posterior samples from lm_b object}
\usage{
get_posterior_draws(object, n_draws = 10000)
}
\arguments{
\item{object}{Object of class lm_b}

\item{n_draws}{integer.  Number of posterior draws to obtain.}
}
\description{
Get posterior samples from lm_b object
}
