% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC.R, R/SDratio.R
\name{IC}
\alias{IC}
\alias{DIC}
\alias{AIC}
\alias{BIC}
\alias{BIC.lm_b}
\alias{AIC.lm_b}
\alias{DIC.lm_b}
\alias{BIC.aov_b}
\alias{AIC.aov_b}
\alias{DIC.aov_b}
\alias{WAIC.lm_b}
\alias{WAIC.aov_b}
\alias{SDratio.lm_b}
\title{Compute AIC, BIC, DIC, or WAIC for aov_b or lm_b objects.  (Lower is better.)}
\usage{
DIC(object)

\method{BIC}{lm_b}(object)

\method{AIC}{lm_b}(object)

\method{DIC}{lm_b}(object, seed = 1)

\method{BIC}{aov_b}(object)

\method{AIC}{aov_b}(object)

\method{DIC}{aov_b}(object)

\method{WAIC}{lm_b}(object, seed = 1)

\method{WAIC}{aov_b}(object)

\method{SDratio}{lm_b}(object)
}
\arguments{
\item{object}{aov_b or lm_b object}

\item{seed}{integer.  Always set your seed!!!}
}
\description{
Compute AIC, BIC, DIC, or WAIC for aov_b or lm_b objects.  (Lower is better.)
}
