% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC.R, R/SDratio.R, R/credint.R
\name{IC}
\alias{IC}
\alias{DIC}
\alias{AIC}
\alias{BIC}
\alias{WAIC}
\alias{BIC.lm_b}
\alias{BIC.glm_b}
\alias{BIC.aov_b}
\alias{AIC.lm_b}
\alias{AIC.glm_b}
\alias{AIC.aov_b}
\alias{DIC.lm_b}
\alias{DIC.glm_b}
\alias{DIC.aov_b}
\alias{WAIC.lm_b}
\alias{WAIC.aov_b}
\alias{WAIC.glm_b}
\alias{SDratio.lm_b}
\alias{credint.lm_b}
\title{Compute AIC, BIC, DIC, or WAIC for aov_b or lm_b objects.  (Lower is better.)}
\usage{
DIC(object, ...)

\method{BIC}{lm_b}(object, ...)

\method{BIC}{glm_b}(object, ...)

\method{BIC}{aov_b}(object, ...)

\method{AIC}{lm_b}(object, ...)

\method{AIC}{glm_b}(object, ...)

\method{AIC}{aov_b}(object, ...)

\method{DIC}{lm_b}(object, seed = 1, mc_relative_error = 0.01, ...)

\method{DIC}{glm_b}(object, seed = 1, ...)

\method{DIC}{aov_b}(object, ...)

\method{WAIC}{lm_b}(object, seed = 1, ...)

\method{WAIC}{aov_b}(object, ...)

\method{WAIC}{glm_b}(object, seed = 1, ...)

\method{SDratio}{lm_b}(object, by = "coefficient", ...)

\method{credint}{lm_b}(object, CI_level = 0.95, ...)
}
\arguments{
\item{object}{aov_b, lm_b, or glm_b object}

\item{...}{optional arguments.
The relative monte carlo error of the expected values of the deviance.
(Ignored for a single population proportion.)}

\item{seed}{integer.  Always set your seed!!!}

\item{mc_relative_error}{The number of posterior draws will ensure that
with 99\% probability the posterior mean of the deviance for DIC will be
within \eqn{\pm}\code{mc_error}E(deviance).}
}
\description{
Compute AIC, BIC, DIC, or WAIC for aov_b or lm_b objects.  (Lower is better.)
}
