% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linear_regression}
\alias{linear_regression}
\title{linear_regression}
\usage{
linear_regression(formula, data)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}}, specifying the model to be fitted.}

\item{data}{A data frame containing the variables referenced in the formula. This should include the response and explanatory variables.}
}
\value{
A matrix containing the estimated beta coefficients and standard errors as well as the t-values, and p-values for each predictor in the model.
}
\description{
A package for performing Ordinary Least Squares (OLS) on a given model and
estimating the corresponding beta coefficients and their standard errors,
as well as conducting pairwise inference via a two-sided t-test.
}
\examples{
# Example of using linear_regression with the mtcars dataset
summary_results <- linear_regression(mpg ~ hp + wt, data = mtcars)
print(summary_results)
}
