% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Title}
\description{
Title
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{B_h}}{vector.}

\item{\code{y_h}}{matrix.}

\item{\code{e}}{matrix.}

\item{\code{df}}{numeric.}

\item{\code{e_var}}{matrix.}

\item{\code{B_h_var}}{vector.}

\item{\code{t_value}}{matrix.}

\item{\code{data_name}}{character.}
}}


