% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_class.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg_class}
\title{Class formatting for RC type object.}
\description{
this list will be filled with objects calculated in \code{linreg()}-function
}

