% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R, R/linreg_class.R, R/linreg_class2.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\alias{linreg_class}
\title{Yep}
\arguments{
\item{formula}{an object of class "formula", a symbolic description of the model to be fitted.}

\item{data}{the data frame including variables to include in the model.}
}
\value{
\code{linreg()} returns a linreg-object with common model-methods like coef, resid, plot etc.
}
\description{
linreg object holds data and methods for data calculated in \code{linreg()}-function.

linreg object holds data and methods for data calculated in \code{linreg()}-function.
}
\section{Fields}{

\describe{
\item{\code{beta_hat}}{Estimates for each beta}

\item{\code{fits}}{Contains fitted values}

\item{\code{residuals}}{Contains residuals, i.e. observed - fitted}

\item{\code{resid_var}}{Residual variance}

\item{\code{beta_info}}{Contains variance, t-value and p-value for each estimated beta value}

\item{\code{df}}{Degrees of freedom for the model}

\item{\code{formula}}{Formula used in model}

\item{\code{call}}{Call sent to \code{linreg()}}

\item{\code{beta_hat}}{Estimates for each beta}

\item{\code{fits}}{Contains fitted values}

\item{\code{residuals}}{Contains residuals, i.e. observed - fitted}

\item{\code{resid_var}}{Residual variance}

\item{\code{beta_info}}{Contains variance, t-value and p-value for each estimated beta value}

\item{\code{df}}{Degrees of freedom for the model}

\item{\code{formula}}{Formula used in model}

\item{\code{call}}{Call sent to \code{linreg()}}
}}


