% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_class.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear Regression Reference Class}
\usage{
linreg(formula, data, beta, predictions, residuals, df, res_var, beta_var, t_value, p_value)
}
\description{
Definition of the Reference Class linreg for the linear regression model
in Lab 4.
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula for building the model}

\item{\code{data}}{dataset that contains feature and target values}

\item{\code{beta}}{values of the model's coefficients}

\item{\code{predictions}}{values of the dependent variable predicted by the model}

\item{\code{residuals}}{values of the residuals}

\item{\code{df}}{degrees of freedom}

\item{\code{res_var}}{variance of the residuals}

\item{\code{beta_var}}{variance of the regression coefficients}

\item{\code{t_value}}{t-values for each coefficient}

\item{\code{p_value}}{p-values for each coefficient}
}}


