% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linreg Function}
\arguments{
\item{formula, }{writen as a formula-object.}

\item{data, }{a data.frame object in this particular case data (iris).}
}
\value{
a linreg class object containing regression outputs.
}
\description{
This function store output from a Regression model in a linreg object.
}
\details{
This function calculates ordrinary least square models (OLS)
}

\examples{
data(iris)
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$print()
}
\author{
Thi Pham
}
