\name{linreg-package}
\alias{linreg-package}
\alias{linreg}
\docType{package}
\title{
\packageTitle{linreg}
}
\description{
\packageDescription{linreg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{linreg}
\packageIndices{linreg}
linear regression
}
\author{
\packageAuthor{linreg}

Maintainer: \packageMaintainer{linreg}
}
\references{

}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
