% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Yep}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{an object of class "formula", a symbolic description of the model to be fitted.}

\item{data}{the data frame including variables to include in the model.}
}
\value{
\code{linreg()} returns a linreg-object with common model-methods like coef, resid, plot etc.
}
\description{
Yep
}
