% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linreg.R
\name{linreg}
\alias{linreg}
\title{Store output from a OLS model in a linreg object}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula, }{writen as a formula-object, see example.}

\item{data, }{a data.frame object}
}
\value{
a linreg class-object containing regression outputs
}
\description{
This function calculates fitted values, residuals, degrees of freedom, t-ratio and p-values
}
\details{
This function calculates ordrinary least square models (OLS)
}
\examples{
linreg(Sepal.Length ~ Sepal.Width, iris)
}
\author{
Yumeng Li, Mattias Karlsson, Ashraf Sarhan
}
