% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression.}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A formula.}

\item{data}{A data frame.}
}
\value{
An object of class linreg.
}
\description{
Linear Regression.
}
