% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression.}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A formula that containing the variables of given dataframe .}

\item{data}{A dataframe.}
}
\value{
An linreg object of Linear regression in s3 class.
}
\description{
Linear Regression.
}
