% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{QR decoposition linear Regression.}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A variable containing in the dataframe.}

\item{data}{A data frame}
}
\value{
An S3 object of class linreg with QR decoposition
}
\description{
QR decoposition linear Regression.
}
