% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegClass.R
\name{LinReg}
\alias{LinReg}
\title{Fit a Linear Model with Truncated Gaussian Error Term}
\usage{
LinReg(formula, data, isLogTransformed = F, constant = 1)
}
\arguments{
\item{formula}{a formula (e.g. "y ~ x")}

\item{data}{a data.frame object}

\item{isLogTransformed}{a logical, it is assumed the response variable has been
log transformed}

\item{constant}{a constant that has been added to the response variable before
the log transformation. It is assumed that this constant is 1.}
}
\value{
an instance of the S3 LinReg class
}
\description{
The function first fits a regular linear model. The parameter
estimates are then used as starting values for the regression
with truncated Gaussian error term.
}
