% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear regression}
\usage{
linreg(formula, data, QR = FALSE)
}
\arguments{
\item{formula}{a formula, like y~x}

\item{data}{a data frame}

\item{QR}{use QR decomposition if TRUE, else OLS}
}
\value{
A linreg class object
}
\description{
The function \code{linreg} takes a formula object and a 
\code{data.frame} as arguments, and returns an object as an S3 class. 
It performs ordinary least squares or QR decomposition to calculate different statistics.
}
\examples{
data(mtcars)

# Fitting a linear regression model using OLS
model_ols <- linreg(formula = mpg ~ wt + cyl, data = mtcars, QR = FALSE)

# Fitting a linear regression model using QR decomposition 
model_qr <- linreg(formula = mpg ~ wt + cyl, data = mtcars, QR = TRUE)

model_ols
model_qr
}
\references{
\url{https://en.wikipedia.org/wiki/QR_decomposition}
}
