% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear regression}
\usage{
linreg(formula, data, QR = FALSE)
}
\arguments{
\item{formula}{a formula, like y~x}

\item{data}{a data frame}

\item{QR}{use QR decomposition if TRUE, else OLS}
}
\value{

}
\description{
Linear regression
}
\examples{
data(mtcars)
model <- linreg(mpg~wt+cyl, mtcars, QR = FALSE)
model
}
