% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{MULTIPLE LINEAR REGRESSION WITH OLS}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{is used to specify which regression model you wish to estimate.}

\item{data}{should be the data frame which you want to estimate the parameters from.}
}
\value{
The object \code{linreg} is returned in the last lines of codes. This includes
the estimated parameters such as the coefficients, fitted values, residuals, degrees of freedom
residual variance, variance of the coefficients, tvalues and pvalues.
}
\description{
This function is used to estimate a multiple linear regression model using the
OLS method for estimating the model parameters.
}
