% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression}
\usage{
linreg(y, x, dataset)
}
\arguments{
\item{y}{outcome or dependent variable}

\item{x}{matrix of covariates or independent variables}

\item{dataset}{dataset containing y and x}
}
\value{
a table of results
}
\description{
Runs linear regression model
}
\examples{
linreg(am, mpg, mtcars)
}
