% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression}
\usage{
linreg(y, x, dataset = NULL)
}
\arguments{
\item{y}{a numeric vector or name of a column vector in the specified dataset}

\item{x}{a list of one or more numeric vectors or names of a column vector in the specified dataset}

\item{dataset}{(optional) a dataset containing the column vectors listed}
}
\value{
The function outputs a table of beta estimates, SE, t statistics, and p values.
}
\description{
This functions fits a linear regression model and provides parameter estimates and significance.
}
\examples{
linreg(mtcars$mpg, list(cyl = mtcars$cyl, disp = mtcars$disp))
linreg("mpg", list(cyl = "cyl", disp = mtcars$dis), mtcars)
}
