% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{liu_col}
\alias{liu_col}
\title{Colour palette function, simply returning LiU colours}
\usage{
liu_col()
}
\value{
A vector of hex codes for LiU's official colors
}
\description{
Colour palette function, simply returning LiU colours
}
\examples{
liu_col()
}
