% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate_b.R
\name{mediate_b}
\alias{mediate_b}
\title{Mediation using Bayesian methods}
\usage{
mediate_b(
  model_m,
  model_y,
  sims = 1000,
  treat,
  control_value,
  treat_value,
  CI_level = 0.95,
  simple = TRUE,
  show_progress = TRUE
)
}
\arguments{
\item{model_m}{a fitted model object of class lm_b for mediator.}

\item{model_y}{a fitted model object of class lm_b for outcome.}

\item{sims}{number of posterior draws for causal quantities}

\item{treat}{a character string indicating the name of the
treatment variable used in the models.  NOTE: Treatment variable must be
numeric.}

\item{control_value}{value of the treatment variable used as the
control condition. Default is the 1st quintile of the treat variable.}

\item{treat_value}{value of the treatment variable used as the treatment condition.
Default is the 4th quintile of the treat variable.}

\item{CI_level}{numeric. Credible interval level.}

\item{simple}{logical.  Set to TRUE only if there are no interactions.}

\item{show_progress}{logical.}
}
\value{
A list with two elements:
\itemize{
\item summary - Posterior mean and credible intervals for causal quantities
\item counterfactual_draws (if \code{simple = FALSE}) - list with the following elements:
\itemize{
\item E_y_00 - The expected value of the counterfactual y(control_value,M(control_value))
\item E_y_01 - The expected value of the counterfactual y(control_value,M(treat_value))
\item E_y_10 - The expected value of the counterfactual y(treat_value,M(control_value))
\item E_y_11 - The expected value of the counterfactual y(treat_value,M(treat_value))
}
\item control_value, treat_value
}
}
\description{
Mediation analysis done in the framework of Imai, Kosuke, et al.
“A General Approach to Causal Mediation Analysis.” Psychological Methods,
vol. 15, no. 4, 2010, pp. 309–34, https://doi.org/10.1037/a0020761.
}
