% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_glm_b.R
\name{np_glm_b}
\alias{np_glm_b}
\title{Non-parametric linear models}
\usage{
np_glm_b(
  formula,
  data,
  family,
  loss = "selfinformation",
  loss_gradient,
  trials,
  n_draws,
  CI_level = 0.95,
  seed = 1
)
}
\arguments{
\item{formula}{A formula specifying the model.}

\item{data}{A data frame in which the variables specified in the formula
will be found. If missing, the variables are searched for in the standard way.}

\item{family}{A description of the error distribution and link function
to be used in the model. See ?glm for more information.}

\item{loss}{Either "selfinformation", "gls" (for generalized squared error),
or a function that takes in two arguments, the first of which should
be the vector of outcomes and the second should be the expected value of y;
The outcome of the function should be the loss evaluated for each observation.
By default, the self-information loss is used (i.e., the negative log-likelihood).
Note: I really do mean the expected value of y, even for binomial (i.e., n*p).}

\item{loss_gradient}{If loss is a user-defined function (as opposed to
"selfinformation" or "gls"), supplying the gradient to the loss will
speed up the algorithm.}

\item{trials}{Integer vector giving the number of trials for each
observation if family = binomial().}

\item{n_draws}{integer.  Number of posterior draws to obtain.  If left missing,
the large sample approximation will be used.}

\item{CI_level}{numeric. Credible interval level.}

\item{seed}{integer.  Always set your seed!!!}
}
\value{
np_glm_b() returns an object of class "np_glm_b", which behaves as
a list with the following elements:
\itemize{
\item summary - a tibble giving results for regression coefficients.
}
}
\description{
np_glm_b uses general Bayesian inference with loss-likelihood bootstrap.
This is, as implemented here, a Bayesian non-parametric linear models
inferential engine.
}
\details{
Applicable data types are continuous (use family =
gaussian()), count (use family = poisson()), or binomial
(use family = binomial()).
}
