% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_class2.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot linreg object}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{A linreg object}

\item{...}{Added for consistency with the default plot}
}
\value{
A residual plot (ggplot2 object)
}
\description{
Plot linreg object
}
