% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot linreg class}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{

}
\description{
Plot linreg class
}
\examples{
data(mtcars)
model <- linreg(mpg~wt+cyl, mtcars)
plot(model)
}
