% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot method for linreg objects}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class "linreg"}

\item{...}{Additional arguments (not used)}
}
\description{
Plot method for linreg objects
}
