% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{plot}
\alias{plot}
\title{Plots}
\usage{
plot(x)
}
\arguments{
\item{x}{An object of class linreg.}
}
\value{
Residual plots.
}
\description{
\code{plot.linreg} Prints residual plots.
}
\details{
This function plots residual plots.
}
