% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.lm_b}
\alias{plot.aov_b}
\alias{plot.lm_b_bma}
\alias{plot.glm_b}
\alias{plot.np_glm_b}
\title{Plots linreg objects.}
\usage{
\method{plot}{lm_b}(
  x,
  type = c("diagnostics", "pdp", "ci band", "pi band"),
  variable,
  exemplar_covariates,
  combine_pi_ci = TRUE,
  variable_seq_length = 30,
  return_as_list = FALSE,
  CI_level = 0.95,
  PI_level = 0.95,
  ...
)

\method{plot}{aov_b}(
  x,
  type = c("diagnostics", "ci band", "pi band"),
  combine_pi_ci = TRUE,
  return_as_list = FALSE,
  CI_level = 0.95,
  PI_level = 0.95,
  ...
)

\method{plot}{lm_b_bma}(
  x,
  type = c("diagnostics", "pdp", "ci band", "pi band"),
  variable,
  exemplar_covariates,
  combine_pi_ci = TRUE,
  bayes_pvalues_quantiles = c(0.01, 1:19/20, 0.99),
  variable_seq_length = 30,
  return_as_list = FALSE,
  CI_level = 0.95,
  PI_level = 0.95,
  seed = 1,
  ...
)

\method{plot}{glm_b}(
  x,
  type,
  variable,
  exemplar_covariates,
  combine_pi_ci = TRUE,
  variable_seq_length = 30,
  return_as_list = FALSE,
  CI_level = 0.95,
  PI_level = 0.95,
  seed = 1,
  ...
)

\method{plot}{np_glm_b}(
  x,
  type,
  variable,
  exemplar_covariates,
  variable_seq_length = 30,
  return_as_list = FALSE,
  CI_level = 0.95,
  seed = 1,
  ...
)
}
\arguments{
\item{x}{A linreg object}

\item{type}{character. Select any of "diagnostics" ("dx" is also allowed),
"pdp" (partial dependence plot), "ci band", and/or "pi band".  NOTE: the CI
and PI bands only work for numeric variables.}

\item{variable}{character. If type = "pdp" , which variable should be plotted?}

\item{exemplar_covariates}{data.frame or tibble with exactly one row.
Used to fix other covariates while varying the variable of interest for the plot.}

\item{combine_pi_ci}{logical. If type includes both "ci band" and "pi band",
should the credible band be superimposed on the prediction band or
plotted separately?}

\item{variable_seq_length}{integer. Number of points used to draw pdp.}

\item{return_as_list}{logical.  If TRUE, a list of ggplots will be returned,
rather than a single plot produced by the patchwork package.}

\item{CI_level}{Posterior probability covered by credible interval}

\item{PI_level}{Posterior probability covered by prediction interval}

\item{...}{optional arguments.}

\item{bayes_pvalues_quantiles}{ADD description!}

\item{seed}{ADD description!}
}
\description{
Plots linreg objects.
}
