% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred}
\alias{pred}
\title{Fitted values}
\usage{
pred(x)
}
\arguments{
\item{x}{An object of class linreg.}
}
\value{
Predicted values yhat.
}
\description{
\code{pred.linreg} Returns the predicted values.
}
\details{
This function returns the predicted values stored in 
the object linreg of S3 class.
}
