% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred}
\alias{pred}
\title{Fitted values}
\usage{
pred(x)
}
\arguments{
\item{x}{A class object of Linear Regression}
}
\value{
A predicted values yhat.
}
\description{
\code{pred.linreg} Returns the predicted values yhat
}
\details{
This function returns  the predicted values y stored in 
the object linreg of S3 class.
}
