% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pred}
\alias{pred}
\title{Prediction caller}
\usage{
pred(x, ...)
}
\arguments{
\item{x}{a linreg class object}

\item{...}{optional arguments to pass to generic}
}
\value{

}
\description{
Prediction caller
}
\examples{
data(mtcars)
model <- linreg(mpg~wt+cyl, mtcars)
pred(model)
}
