% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.aov_b.R
\name{predict.aov_b}
\alias{predict.aov_b}
\title{Predict method for aov_b model fits}
\usage{
\method{predict}{aov_b}(object, CI_level = 0.95, PI_level = 0.95)
}
\arguments{
\item{object}{Object of class aov_b}

\item{CI_level}{Posterior probability covered by credible interval}

\item{PI_level}{Posterior probability covered by prediction interval}
}
\value{
tibble with estimate (posterior mean), prediction intervals, and credible intervals
for the mean.
}
\description{
Predict method for aov_b model fits
}
