% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_class2.R
\name{predict.linreg}
\alias{predict.linreg}
\title{Predictions from linreg object}
\usage{
\method{predict}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg object}

\item{...}{Additional arguments}
}
\value{
Numeric vector of predictions
}
\description{
Predictions from linreg object
}
