% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegClass.R
\name{predict.LinReg}
\alias{predict.LinReg}
\title{Provide Predictions on the Original Scale.}
\usage{
\method{predict}{LinReg}(object, newdata, estimator)
}
\arguments{
\item{object}{a LinReg S3 instance}

\item{newdata}{a data.frame object with all the fields required by the model plus the y field set to NA}

\item{estimator}{a character string among these ones: Baskerville, BeauchampAndOlson, MonteCarlo, or ComplexMonteCarlo}
}
\description{
The method assumes that the responses are log transformed.
}
