% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.linreg}
\alias{predict.linreg}
\title{Predictions for linreg class}
\usage{
\method{predict}{linreg}(object, ...)
}
\arguments{
\item{object}{}

\item{...}{}
}
\value{

}
\description{
Predictions for linreg class
}
