% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.linreg}
\alias{predict.linreg}
\title{Predictions for linreg class}
\usage{
\method{predict}{linreg}(object, ...)
}
\arguments{
\item{object}{a linreg class object}

\item{...}{optional arguments to pass to generic}
}
\value{

}
\description{
Predictions for linreg class
}
\examples{
data(mtcars)
model <- linreg(mpg~wt+cyl, mtcars)
pred(model)
}
