% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_class2.R
\name{print.linreg}
\alias{print.linreg}
\title{Print linreg object}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{A linreg object}

\item{...}{Additional arguments}
}
\value{
Prints the function call and regression coefficients
}
\description{
Print linreg object
}
