% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.linreg}
\alias{print.linreg}
\title{Print for linreg class}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{a linreg class object}

\item{...}{optional arguments to pass to generic}
}
\value{

}
\description{
Print for linreg class
}
\examples{
data(mtcars)
model <- linreg(mpg~wt+cyl, mtcars)
print(model)
}
