% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{print}
\alias{print}
\title{Print}
\usage{
print(x)
}
\arguments{
\item{x}{A class object of Linear Regression}
}
\value{
the coefficients and coefficient names
}
\description{
\code{print.linreg} print out the coefficients and coefficient names
}
\details{
This function returns the coefficients and coefficient names stored in 
the object linreg of S3 class.
}
