% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.aov_b}
\alias{print.lm_b}
\alias{print.np_lm_b}
\title{Print linreg objects.}
\usage{
print.aov_b(object)

print.lm_b(object)

print.np_lm_b(object)
}
\description{
Print linreg objects.
}
