% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.aov_b}
\alias{print.lm_b}
\alias{print.np_glm_b}
\alias{print.lm_b_bma}
\alias{print.glm_b}
\alias{print.mediate_b}
\title{Print linreg objects.}
\usage{
\method{print}{aov_b}(x, ...)

\method{print}{lm_b}(x, ...)

\method{print}{np_glm_b}(x, ...)

\method{print}{lm_b_bma}(x, ...)

\method{print}{glm_b}(x, ...)

\method{print}{mediate_b}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{optional arguments.}
}
\description{
Print linreg objects.
}
