% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.aov_b}
\alias{print.lm_b}
\alias{print.np_lm_b}
\alias{print.bma}
\title{Print linreg objects.}
\usage{
print.aov_b(x)

print.lm_b(x)

print.np_lm_b(x)

print.bma(x)
}
\arguments{
\item{x}{an object used to select a method.}
}
\description{
Print linreg objects.
}
