% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_class2.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Extract residuals from linrej object}
\usage{
\method{residuals}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg object}

\item{...}{Additional arguments}
}
\value{
Numeric vector of residuals
}
\description{
Extract residuals from linrej object
}
