% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegClass.R
\name{residuals.LinReg}
\alias{residuals.LinReg}
\title{Provide the Model Residuals}
\usage{
\method{residuals}{LinReg}(object)
}
\arguments{
\item{object}{an instance of the S3 class LinReg}
}
\value{
a vector of residuals
}
\description{
Provide the Model Residuals
}
