% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Residuals for linreg class}
\usage{
\method{residuals}{linreg}(object, ...)
}
\arguments{
\item{object}{a linreg class object}

\item{...}{optional arguments to pass to generic}
}
\value{
A vector of residuals for an linreg class object
}
\description{
Residuals for linreg class
}
\examples{
data(mtcars)
model <- linreg(mpg~wt+cyl, mtcars, QR = FALSE)
resid(model)
}
