% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Residuals for linreg class}
\usage{
\method{residuals}{linreg}(object, ...)
}
\arguments{
\item{object}{}

\item{...}{}
}
\value{

}
\description{
Residuals for linreg class
}
\examples{
data(mtcars)
model <- linreg(mpg~wt+cyl, mtcars)
resid(model)
}
