% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Extract residuals from linreg objects (alternative name)}
\usage{
\method{residuals}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class "linreg"}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of residuals
}
\description{
Extract residuals from linreg objects (alternative name)
}
