% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Ridge Regression Reference Class}
\value{
An instance of the `ridgereg` class with fields and methods for ridge regression.
}
\description{
This reference class implements ridge regression, allowing users to fit a ridge regression model, compute coefficients, predict values, and print model summaries.
}
\details{
Ridge regression is useful in cases where multicollinearity is present among predictors or when the number of predictors (p) exceeds the number of observations (n). This class calculates the coefficients using the formula:
\deqn{\beta_{ridge} = (X^TX + \lambda I)^{-1}X^Ty}
where \eqn{\lambda} is the regularization parameter.
}
\section{Fields}{

\describe{
\item{\code{formula}}{A formula object specifying the regression model.}

\item{\code{data_name}}{A character string with the name of the data used in the model.}

\item{\code{data}}{A data frame containing the dataset to be used for ridge regression.}

\item{\code{lambda}}{A numeric value for the ridge penalty parameter, \eqn{\lambda}.}

\item{\code{coefficients}}{A numeric vector storing the ridge regression coefficients.}

\item{\code{fitted_values}}{A numeric vector of the model's fitted values.}
}}


