% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Ridge Regression Reference Class}
\arguments{
\item{formula}{A formula specifying the model to be used.}

\item{data}{A data frame containing the data for fitting the model.}

\item{data_name}{A character string specifying the name of the dataset.}

\item{lambda}{A numeric value for the ridge penalty parameter (default is 0).}

\item{newdata}{Optional data frame for predicting new values. If NULL, the method returns fitted values for the training data.}
}
\value{
An instance of the `ridgereg` class with fields and methods for ridge regression.

A numeric vector of predicted values.

A numeric vector of ridge regression coefficients.
}
\description{
This reference class implements ridge regression, allowing users to fit a ridge regression model, compute coefficients, predict values, and print model summaries.
}
\details{
Ridge regression is useful in cases where multicollinearity is present among predictors or when the number of predictors (p) exceeds the number of observations (n). This class calculates the coefficients using the formula:
\deqn{\beta_{ridge} = (X^TX + \lambda I)^{-1}X^Ty}
where \eqn{\lambda} is the regularization parameter.
}
\section{Fields}{

\describe{
\item{\code{formula}}{A formula object specifying the regression model.}

\item{\code{data_name}}{A character string with the name of the data used in the model.}

\item{\code{data}}{A data frame containing the dataset to be used for ridge regression.}

\item{\code{lambda}}{A numeric value for the ridge penalty parameter, \eqn{\lambda}.}

\item{\code{coefficients}}{A numeric vector storing the ridge regression coefficients.}

\item{\code{fitted_values}}{A numeric vector of the model's fitted values.}
}}


