% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{Store output from a OLS model in a ridgereg object}
\usage{
ridgereg(formula, data, lambda)
}
\arguments{
\item{formula, }{writen as a formula-object, see example.}

\item{data, }{a data.frame object with standardized variables}

\item{lambda, }{hyperparameter used to find the best model}
}
\value{
a ridgereg class-object containing regression outputs
}
\description{
This function calculates fitted values, residuals, degrees of freedom, t-ratio and p-values
}
\details{
This function calculates Ridge regression
}
\author{
Yumeng Li, Mattias Karlsson, Ashraf Sarhan
}

