% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ridgereg.R
\docType{class}
\name{ridgereg}
\alias{ridgereg}
\title{Ridge Regression Class}
\arguments{
\item{formula}{A formula object}

\item{data}{A dataset}
}
\value{
Returns a ridge model built based on \code{formula} and \code{data}
}
\description{
This class is implemented to handle special output
    functions such as print(), plot(), resid(), pred(), coef()
    and summary(). Please refer the methods section of documentation
    for information on each method
}
\details{
Advanced Programming in R - Bonus Lab
Ridge Regression using ordinary least squares
}
\section{Methods}{

\describe{
\item{\code{coef()}}{Return the ridge regression coefficients as a named vector}

\item{\code{predict()}}{Return the predicted values}

\item{\code{print()}}{Print out the ridge regression coefficients and coefficient names}

\item{\code{resid()}}{Return the vector of residuals}
}}

\examples{
\dontrun{
obj = ridgereg(formula = Petal.Length ~ Species, data = iris)
obj$print()
obj$plot()
}
}
