% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_class2.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary of linreg object}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg object}

\item{...}{Additional arguments}
}
\value{
Prints coefs, se(coefs), t-value, p-value.
}
\description{
Summary of linreg object
}
