% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegClass.R
\name{summary.LinReg}
\alias{summary.LinReg}
\title{Provide a Summary of the fit}
\usage{
\method{summary}{LinReg}(object, ...)
}
\arguments{
\item{object}{an instance of the S3 class LinReg}

\item{...}{absolutely useless, just to keep the signature consistent for inheritance}
}
\description{
Provide a Summary of the fit
}
