% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary for linreg class}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{a linreg class object}

\item{...}{optional arguments to pass to generic}
}
\value{

}
\description{
Summary for linreg class
}
\examples{
data(mtcars)
model <- linreg(mpg~wt+cyl, mtcars)
summary(model)
}
