% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summarize a linreg model}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns an object of class \code{summary.linreg}.
}
\description{
Creates a summary containing the estimated coefficients, standard errors,
t-values, p-values, residual standard error and degrees of freedom.
}
