% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary method for linreg objects}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class "linreg"}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for linreg objects
}
