% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{summary}
\alias{summary}
\title{Regression summary}
\usage{
summary(x)
}
\arguments{
\item{x}{An object of class linreg.}
}
\value{
Summary of calculated values.
}
\description{
\code{summary.linreg} Outputs a summary of the calculated results from the regression model.
}
