% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.lm_b}
\alias{summary.aov_b}
\alias{summary.np_lm_b}
\title{Summary functions for lm_b and aov_b objects}
\usage{
summary.lm_b(object, CI_level = 0.95)

summary.aov_b(object, CI_level = 0.95)

summary.np_lm_b(object)
}
\arguments{
\item{object}{lm_b or aov_b object}
}
\description{
Summary functions for lm_b and aov_b objects
}
