% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.lm_b}
\alias{summary.aov_b}
\title{Summary functions for lm_b and aov_b objects}
\usage{
\method{summary}{lm_b}(object)

\method{summary}{aov_b}(object)
}
\arguments{
\item{object}{lm_b or aov_b object}
}
\description{
Summary functions for lm_b and aov_b objects
}
