% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.lm_b}
\alias{summary.aov_b}
\alias{summary.np_glm_b}
\alias{summary.lm_b_bma}
\alias{summary.glm_b}
\title{Summary functions for linreg objects}
\usage{
summary.lm_b(object, CI_level = 0.95)

summary.aov_b(object, CI_level = 0.95)

summary.np_glm_b(object, CI_level = 0.95, interpretable_scale = TRUE)

summary.lm_b_bma(object, CI_level = 0.95)

summary.glm_b(object, CI_level = 0.95, interpretable_scale = TRUE)
}
\arguments{
\item{object}{linreg object}

\item{CI_level}{Posterior probability covered by credible interval}
}
\description{
Summary functions for linreg objects
}
