% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{theme_liu}
\alias{theme_liu}
\title{Theme with LiU colours}
\usage{
theme_liu()
}
\value{

}
\description{
Theme with LiU colours
}
\examples{
data(mtcars)
lm <- linreg(mpg~wt+cyl, data = mtcars)
plot(lm)
}
